feat(translator): deprecate the built-in OpenAI adapter, and say what the schema envelope costs
The adapter exists to carry a dependency that is not ours, and carrying it
*optionally* is what makes it expensive: `loadOpenAIClient` lazily imports the SDK
through a module-level constant because deployment file-tracers resolve `import()`
statically and silently prune anything else, and `OpenAI.shapes.ts` hand-writes a
structural slice of the client, with a conformance test to catch drift, so no
vendor type reaches the emitted declarations. A consumer importing the SDK they
already chose needs none of that.

The 0.11.0 layering is what makes removal possible. Once
`createTranslationProvider({ complete })` exists, the wrapper's remaining job is
defaults and a request body — about a dozen lines, now a recipe in the README, and
the SDK version becomes the consumer's. `createOpenAIProvider` and
`OpenAIProviderConfig` are annotated; the ledger entry names the whole
`openai/` directory and lists the rest of the surface, along with what a consumer
gives up: the message naming `structuredOutput` when a gateway rejects a strict
schema, the guard against `json_object` with a prompt that never says "json", and
the 60 s timeout in place of the SDK's ten minutes. It also settles the shape of
issue #100 — the package will ship recipes, not three more adapters.

**`structuredOutput` now documents both of its risks.** It carried one: some
gateways reject a strict schema. The second was discovered after that text was
written and is the more surprising: the schema has a size limit, so one request
carries a limited number of pieces of text and a document past that ceiling fails
as a whole. The count is ours to explain — one property per translatable piece,
and rich text splits one piece per text node, so a sentence with two emphasised
spans is already four — and nothing splits a document across requests. Presented
as a choice between two risks rather than a right mode and a fallback: the
alternative has no ceiling but only *detects* a dropped key, and the report
reaches a server log no editor sees.

Consumer-facing text states consequences only. The vendor's mechanism, the list of
models that rejected the strict envelope as of today, and an unverified claim about
when billing applies all moved to the design doc, where going stale costs nothing.

Also here:

- The deprecated `OpenAITranslationProvider` class moves to its own file with its
  tests beside it, so the next-major removal is a deletion rather than surgery.
  `.deprecated.ts` joins the role-tag vocabulary, with the rule that such a file
  holds only deprecated surface — a directory listing then shows what dies next
  major without opening the ledger.
- One dangling ledger reference fixed; it pointed at the path the provider layer
  moved out of. All 13 code refs now resolve, checked by script.
- `createTranslationProvider`'s `@example` no longer shows `signal` wired into a
  service call. The field is reserved and always `undefined`; the example was
  public documentation implying otherwise.

No behaviour change: no default moved, every export keeps its name and its path,
and `src/index.ts` is untouched. 1291 tests across 107 files, check-types clean,
lint 0 errors and 0 warnings on every touched file.

Refs #99, #100
