+++ packages/payload-plugin-translator/docs/DEPRECATIONS.md
@@ -106,7 +106,7 @@ the single source of truth — code annotations link here by anchor instead of d
 - **Deprecated:** pre-ledger
 - **Replacement:** `createOpenAIProvider()` factory.
 - **Remove in:** next major
-- **Code refs:** `src/server/modules/translation-providers/OpenAITranslation.provider.ts`
+- **Code refs:** `src/translation-providers/openai/OpenAITranslationProvider.deprecated.ts`
 
 ### provider-dry-run
 
@@ -138,3 +138,36 @@ the single source of truth — code annotations link here by anchor instead of d
 - **Code refs:** `src/translation-providers/shared/runDryRun.ts`,
   `src/translation-providers/shared/CompletionProvider.provider.ts`,
   `src/translation-providers/openai/OpenAITranslation.provider.ts`
+
+### built-in-openai-adapter
+
+- **What:** the whole built-in OpenAI adapter — `createOpenAIProvider`, `OpenAIProviderConfig`,
+  `OpenAIClientShape`, `OpenAISamplingParams`, `OpenAIStructuredOutput`, and the already-deprecated
+  `OpenAITranslationProvider` class that delegates to it. Everything under
+  `src/translation-providers/openai/` goes as a unit, and the `openai` optional dependency goes with
+  it.
+- **Status:** live
+- **Deprecated:** 2026-08-28 / PR #101
+- **Replacement:** `createTranslationProvider({ complete })` plus your own `openai` client — about a
+  dozen lines, recipe in the README. The SDK version becomes yours, and the envelope choice sits in
+  your own code where you can see it.
+- **Remove in:** next major
+- **Why:** the adapter exists to carry someone else's dependency, and carrying it optionally is what
+  makes it expensive. `loadOpenAIClient` lazily imports the SDK through a module-level constant
+  because deployment file-tracers resolve `import()` statically and silently prune anything else;
+  `OpenAI.shapes.ts` hand-writes a structural slice of the SDK's client, with a conformance test to
+  catch drift, so no vendor type enters the emitted declarations. None of that machinery is needed
+  by a consumer who simply imports the SDK they already chose.
+
+  The 0.11.0 layering made this possible: once `createTranslationProvider({ complete })` exists and
+  `createOpenAIProvider` accepts an injected `client`, the wrapper's remaining job is defaults and a
+  request body — a dozen lines the consumer is better placed to own.
+
+  What a consumer gives up, stated plainly so the README recipe can cover it: the message that names
+  `structuredOutput: "json_object"` when a gateway rejects a strict schema, the guard against
+  `json_object` with a prompt that never says "json", and the 60 s default timeout in place of the
+  SDK's ten minutes.
+
+  This also settles the shape of the planned Anthropic / Gemini / OpenRouter work (issue #100): the
+  package ships recipes, not adapters.
+- **Code refs:** `src/translation-providers/openai/` (the directory as a whole)
