+++ packages/payload-plugin-translator/src/translation-providers/openai/OpenAITranslation.provider.test.ts
@@ -7,7 +7,7 @@ import {
   TransportError,
   UnparseableReplyError,
 } from "../shared";
-import { createOpenAIProvider, OpenAITranslationProvider } from "./OpenAITranslation.provider";
+import { createOpenAIProvider } from "./OpenAITranslation.provider";
 import type { OpenAIChatParams, OpenAIClientShape } from "./OpenAI.shapes";
 import { loadOpenAIClient } from "./loadOpenAIClient";
 
@@ -469,39 +469,5 @@ describe("createOpenAIProvider", () => {
 
       expect(build).toBeTypeOf("function");
     });
-  });
-});
-
-describe("OpenAITranslationProvider (deprecated class)", () => {
-  beforeEach(() => {
-    vi.clearAllMocks();
-  });
-
-  it("still works, and produces what the factory produces", async () => {
-    const { client } = stubClient('{"0":"Hallo"}');
-
-    const fromClass = await new OpenAITranslationProvider({ client }).translate(
-      { 0: "Hello" },
-      "en",
-      "de"
-    );
-    const fromFactory = await createOpenAIProvider({ client }).translate(
-      { 0: "Hello" },
-      "en",
-      "de"
-    );
-
-    expect(fromClass).toEqual(fromFactory);
-  });
-
-  it("delegates rather than re-implementing — a dry run short-circuits there too", async () => {
-    const result = await new OpenAITranslationProvider({ apiKey: "", dryRun: true }).translate(
-      { 0: "Hello" },
-      "en",
-      "de"
-    );
-
-    expect(result).toEqual({ 0: "olleH" });
-    expect(loadClient).not.toHaveBeenCalled();
   });
 });
